# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Results API","description":"REST API for managing match results","contact":{"name":"Juan Francisco Fernandez Herreros","email":"juanfranciscofernandezherreros@gmail.com"},"license":{"name":"Unlicensed"},"version":"0.1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Results","description":"Operations for managing match results"}],"paths":{"/api/results/{id}":{"get":{"tags":["Results"],"summary":"Get a result by id","description":"Returns the match result associated with the specified identifier.","operationId":"getResultById","parameters":[{"name":"id","in":"path","description":"Unique identifier of the result","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Result found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultResponseDTO"}}}},"404":{"description":"Result not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"put":{"tags":["Results"],"summary":"Update a result","description":"Fully updates an existing match result.","operationId":"updateResult","parameters":[{"name":"id","in":"path","description":"Unique identifier of the result to update","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultRequestDTO"}}},"required":true},"responses":{"200":{"description":"Result updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultResponseDTO"}}}},"400":{"description":"Invalid request payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Result not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["Results"],"summary":"Delete a result","description":"Deletes the match result associated with the specified identifier.","operationId":"deleteResult","parameters":[{"name":"id","in":"path","description":"Unique identifier of the result to delete","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"204":{"description":"Result deleted"},"404":{"description":"Result not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/results":{"get":{"tags":["Results"],"summary":"Get a paginated list of results","description":"Returns match results matching the given filters (all optional), paginated and sortable. String filters match partially and case-insensitively; id, score, and date filters match exactly.","operationId":"getAllResults","parameters":[{"name":"id","in":"query","description":"Exact result identifier","required":false,"schema":{"type":"integer","format":"int64","description":"Exact result identifier","example":1},"example":1},{"name":"homeTeam","in":"query","description":"Partial, case-insensitive match on the home team name","required":false,"schema":{"type":"string","description":"Partial, case-insensitive match on the home team name","example":"Real"},"example":"Real"},{"name":"awayTeam","in":"query","description":"Partial, case-insensitive match on the away team name","required":false,"schema":{"type":"string","description":"Partial, case-insensitive match on the away team name","example":"Barcelona"},"example":"Barcelona"},{"name":"homeScore","in":"query","description":"Exact number of goals scored by the home team","required":false,"schema":{"type":"integer","format":"int32","description":"Exact number of goals scored by the home team","example":2},"example":2},{"name":"awayScore","in":"query","description":"Exact number of goals scored by the away team","required":false,"schema":{"type":"integer","format":"int32","description":"Exact number of goals scored by the away team","example":1},"example":1},{"name":"matchDate","in":"query","description":"Exact match date","required":false,"schema":{"type":"string","format":"date","description":"Exact match date","example":"2026-02-05"},"example":"2026-02-05"},{"name":"competition","in":"query","description":"Partial, case-insensitive match on the competition name","required":false,"schema":{"type":"string","description":"Partial, case-insensitive match on the competition name","example":"La Liga"},"example":"La Liga"},{"name":"venue","in":"query","description":"Partial, case-insensitive match on the venue name","required":false,"schema":{"type":"string","description":"Partial, case-insensitive match on the venue name","example":"Bernabeu"},"example":"Bernabeu"},{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","default":["id,ASC"],"items":{"type":"string"}}}],"responses":{"200":{"description":"Results retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PagedModel"}}}}}},"post":{"tags":["Results"],"summary":"Create a result","description":"Creates and persists a new match result.","operationId":"createResult","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultRequestDTO"}}},"required":true},"responses":{"201":{"description":"Result created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultResponseDTO"}}}},"400":{"description":"Invalid request payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"ResultRequestDTO":{"type":"object","description":"Data required to create or update a match result","properties":{"homeTeam":{"type":"string","description":"Name of the home team","example":"Real Madrid","minLength":1},"awayTeam":{"type":"string","description":"Name of the away team","example":"Barcelona","minLength":1},"homeScore":{"type":"integer","format":"int32","description":"Goals scored by the home team","example":2,"minimum":0},"awayScore":{"type":"integer","format":"int32","description":"Goals scored by the away team","example":1,"minimum":0},"matchDate":{"type":"string","format":"date","description":"Date the match was played","example":"2026-02-05"},"competition":{"type":"string","description":"Competition or tournament the match belongs to","example":"La Liga"},"venue":{"type":"string","description":"Venue where the match was played","example":"Santiago Bernabeu"}},"required":["awayScore","awayTeam","homeScore","homeTeam","matchDate"]},"ResultResponseDTO":{"type":"object","description":"Match result returned by the API","properties":{"id":{"type":"integer","format":"int64","description":"Unique identifier generated by the server","example":1,"readOnly":true},"homeTeam":{"type":"string","description":"Name of the home team","example":"Real Madrid"},"awayTeam":{"type":"string","description":"Name of the away team","example":"Barcelona"},"homeScore":{"type":"integer","format":"int32","description":"Goals scored by the home team","example":2},"awayScore":{"type":"integer","format":"int32","description":"Goals scored by the away team","example":1},"matchDate":{"type":"string","format":"date","description":"Date the match was played","example":"2026-02-05"},"competition":{"type":"string","description":"Competition or tournament the match belongs to","example":"La Liga"},"venue":{"type":"string","description":"Venue where the match was played","example":"Santiago Bernabeu"}}},"ErrorResponse":{"type":"object","description":"Standard API error response","properties":{"timestamp":{"type":"string","format":"date-time","description":"Time the error occurred","example":"2026-01-01T12:00:00Z"},"status":{"type":"integer","format":"int32","description":"HTTP status code","example":404},"error":{"type":"string","description":"HTTP status reason phrase","example":"Not Found"},"message":{"type":"string","description":"Human-readable error message","example":"Result not found with id: 999"},"path":{"type":"string","description":"Request path that produced the error","example":"/api/results/999"}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"}}},"PagedModel":{"type":"object","properties":{"content":{"type":"array","items":{}},"page":{"$ref":"#/components/schemas/PageMetadata"}}}}}}